Convert to OOP for reusability
