+++ app/src/main/kotlin/io/github/projectbukirin/palaylab/PredictionCounter.kt
@@ -1,0 +1,39 @@
+package io.github.projectbukirin.palaylab
+
+import android.widget.TextView
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.database.*
+import java.text.DateFormat
+import java.text.SimpleDateFormat
+import java.util.*
+
+class PredictionCounter(val detectionName: String, val textViewId: TextView) {
+    private val firebaseDatabase = FirebaseDatabase.getInstance()
+    private val dbReference = firebaseDatabase.getReference("palaylab-users-prod-01")
+    private val userId: String? = FirebaseAuth.getInstance().currentUser?.uid
+    private val dateFormat: DateFormat = SimpleDateFormat("yyyy-MM-dd")
+    private val date = Date()
+    private val dateToday = dateFormat.format(date)
+    private val query: Query = dbReference.child("predictions").orderByChild("uuid").equalTo(userId)
+
+    fun getPredictionTotal() {
+        query.addValueEventListener(object : ValueEventListener {
+            override fun onDataChange(snapshot: DataSnapshot) {
+                var ctr = 0
+                for (ds in snapshot.children) {
+                    val dName = ds.child("detectionName").getValue<String>(String::class.java)
+                    val dDate = ds.child("detectionDate").getValue<String>(String::class.java)
+                    if (dName == detectionName && dDate == dateToday) {
+                        ctr++
+                    } else {
+                    }
+                }
+                textViewId.text = ctr.toString()
+            }
+
+            override fun onCancelled(error: DatabaseError) {
+                TODO("Not yet implemented")
+            }
+        })
+    }
+}
